requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0

# Trading and financial analysis
ta-lib>=0.4.0  # Technical analysis library (optional, for advanced indicators)

# Data handling
requests>=2.25.0
websockets>=10.0

# Configuration and utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Logging and monitoring
colorlog>=6.0.0

# Dashboard and visualization
streamlit>=1.28.0

# Performance monitoring
psutil>=5.8.0

# Hyperliquid SDK dependencies
eth-utils>=2.1.0,<6.0.0
eth-account>=0.10.0,<0.14.0
websocket-client>=1.5.1
msgpack>=1.0.5

# Development and testing
pytest>=6.0.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
